Revert "Update: Improve TitleOrInfoLog hierarchy & layout"
This reverts commit 95dbceb929c22e84ecb64e33305e5b182dad0a9c.

The code clips the most important info at the bottom in favor of an animation.
The animation takes time, that is not wanted in that situation.
It doesn't work anyways, the text jumps to the small one blow the big one, then rotates. It's not a smooth morphing from one to the other.
Power users don't need that animation and normal users do not use the loginfo feature.